Configure chromatic, improve build flow, add UDLogo component
