+++ src/components/UDLogo/index.ts
@@ -1,0 +1,2 @@
+export {default} from './UDLogo';
+export * from './UDLogo';
