+++ src/components/UDLogoAsRay/UDLogoAsRay.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import type {SvgIconProps} from '@mui/material';
+import {SvgIcon} from '@mui/material';
+
+const UDLogoAsRay = (props: SvgIconProps) => (
+  <SvgIcon viewBox="0 0 94 94" {...props}>
+    <path
+      d="M3.91699 74.0115L3.92238 74.0074L3.93315 73.9993L3.91699 74.0115L3.94796 73.9885L20.1973 61.7646C20.1152 60.9059 20.0732 60.0354 20.0732 59.1552V37.546L36.2295 28.6322V49.7011L55.0784 35.5201V18.2328L73.9274 7.83333V21.3377L90.0837 9.1839V11.8851L73.9274 23.5324V25.7271L90.0837 14.5862V17.2874L73.9274 27.9218V30.1165L90.0837 19.9885V22.6897L73.9274 32.3111V34.5072L90.0837 25.3908V28.0919L73.9274 36.7019V38.8952L90.0837 30.7931V33.4942L73.9274 41.0899V43.2846L90.0837 36.1954V38.8965L73.9274 45.4793V59.1552C73.9274 74.0732 61.8717 86.1667 47.0003 86.1667C34.8349 86.1667 24.5538 78.0738 21.215 66.9622L3.92776 74.0061L3.91699 74.0115ZM20.9462 66.0034L3.93853 73.9993L3.9318 74.0034L21.0752 66.4804C21.0308 66.322 20.9878 66.163 20.9462 66.0034ZM20.3701 63.1798L3.94796 73.9885L3.93315 73.9993L20.4444 63.6499C20.4183 63.4937 20.3935 63.337 20.3701 63.1798ZM20.7183 65.0589L3.96008 73.9885L20.8274 65.5296C20.7897 65.3733 20.7533 65.2164 20.7183 65.0589ZM20.2467 62.238L3.96815 73.975L20.3043 62.7097C20.2838 62.5529 20.2645 62.3956 20.2467 62.238ZM55.0784 53.1613L36.3692 60.7863C37.1397 65.2303 41.0035 68.6092 45.654 68.6092C50.859 68.6092 55.0784 64.3765 55.0784 59.1552V53.1613ZM55.0784 49.9524L36.2295 58.8162V59.1552C36.2295 59.3783 36.2372 59.5995 36.2523 59.8188L55.0784 51.5568V49.9524ZM55.0784 46.7474L36.2295 56.7917V57.8032L55.0784 48.3492V46.7474ZM55.0784 43.5371L36.2295 54.7645V55.7787L55.0784 45.143V43.5371ZM55.0784 40.3295L36.2295 52.7386V53.7515L55.0784 41.934V40.3295ZM55.0784 37.1232L36.2295 50.7141V51.7257L55.0784 38.7264V37.1232ZM3.9318 74.0034L3.92776 74.0061L3.93853 73.9993L20.6181 64.5881C20.5863 64.4318 20.5558 64.275 20.5267 64.1177L3.93315 73.9993L3.92776 74.0061L3.92238 74.0074L3.9318 74.0034Z"
+      fill="white"
+    />
+  </SvgIcon>
+);
+
+export default UDLogoAsRay;
