+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export {default as UDLogo} from './UDLogo';
+export * from './UDLogo';
+
+export {default as UDLogoAsRay} from './UDLogoAsRay';
