+++ src/index.ts
@@ -1,1 +1,3 @@
 export * from '@mui/material';
+
+export * from './components';
